requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
abseil-cpp=20210324.1=h9c3ff4c_0
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py37h5e8e339_0
arrow-cpp=1.0.1=py37haa335b2_40_cuda
arrow-cpp-proc=3.0.0=cuda
async-timeout=3.0.1=py37h06a4308_0
attrs=21.2.0=pyhd8ed1ab_0
aws-c-cal=0.5.11=h95a6274_0
aws-c-common=0.6.2=h7f98852_0
aws-c-event-stream=0.2.7=h3541f99_13
aws-c-io=0.10.5=hfb6a706_0
aws-checksums=0.1.11=ha31a3da_7
aws-sdk-cpp=1.8.186=hb4091e7_3
blas=1.0=mkl
blinker=1.4=py_1
bokeh=2.3.3=py37h89c1867_0
boost-cpp=1.72.0=h9d3c048_4
brotli=1.0.9=h7f98852_5
brotli-bin=1.0.9=h7f98852_5
brotlipy=0.7.0=py37h5e8e339_1001
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h7f98852_1
ca-certificates=2021.10.8=ha878542_0
cachetools=4.2.2=pyhd8ed1ab_0
certifi=2021.10.8=py37h89c1867_1
cffi=1.14.6=py37h400218f_0
chardet=4.0.0=py37h89c1867_2
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=7.1.2=pyh9f0ad1d_0
cloudpickle=2.0.0=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=35.0.0=py37h5d9358c_0
cudatoolkit=10.2.89=hfd86e86_1
cudf=0.19.2=cuda_10.2_py37_gab3b3f653a_0
cudnn=8.2.1.32=h2c0ae14_0
cuml=0.19.0=cuda10.2_py37_g4cb78ff1a_0
cupy=8.6.0=py37hfed0110_0
cutensor=1.2.2.5=h1a5f58c_5
cytoolz=0.11.0=py37h5e8e339_3
dask=2021.4.0=pyhd8ed1ab_0
dask-core=2021.4.0=pyhd8ed1ab_0
dask-cudf=0.19.2=py37_gab3b3f653a_0
dataclasses=0.8=pyhc8e2a94_3
distributed=2021.4.0=py37h89c1867_0
dlpack=0.3=he1b5a44_1
faiss-proc=1.0.0=cuda
fastavro=1.4.4=py37h5e8e339_0
fastrlock=0.6=py37hcd2ae1e_1
ffmpeg=4.2.2=h20bf706_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.8.1=pyhd8ed1ab_0
future=0.18.2=py37h89c1867_4
gflags=2.2.2=he1b5a44_1004
glog=0.5.0=h48cff8f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.3=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpc-cpp=1.38.0=h2519f57_0
grpcio=1.37.1=py37hb27c1af_0
heapdict=1.0.1=py_0
hyperopt=0.1.2=py37h39e3cac_0
icu=68.1=h58526e2_0
idna=3.1=pyhd3deb0d_0
imbalanced-learn=0.8.1=pyhd8ed1ab_0
importlib-metadata=4.8.2=py37h89c1867_0
intel-openmp=2021.3.0=h06a4308_3350
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
krb5=1.19.2=hcc1bbae_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=11_linux64_mkl
libbrotlicommon=1.0.9=h7f98852_5
libbrotlidec=1.0.9=h7f98852_5
libbrotlienc=1.0.9=h7f98852_5
libcblas=3.9.0=11_linux64_mkl
libcudf=0.19.2=cuda10.2_gab3b3f653a_0
libcuml=0.19.0=cuda10.2_g4cb78ff1a_0
libcumlprims=0.19.0=cuda10.2_ga2abf9f_0
libcurl=7.78.0=h2574ce0_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_3
libfaiss=1.7.0=cuda102hd37495c_8_cuda
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=11.2.0=h69a702a_8
libgfortran5=11.2.0=h5c6108e_8
libgomp=9.3.0=h5101ec6_17
libhwloc=2.3.0=h5e5b7d1_1
libiconv=1.16=h516909a_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=11_linux64_mkl
libllvm10=10.0.1=he513fc3_3
libnghttp2=1.43.0=h812cca2_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.16.0=h780b84a_0
librmm=0.19.0=cuda10.2_g7065af3_0
libssh2=1.9.0=ha56f1ee_6
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libthrift=0.14.1=he6d91bd_2
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libutf8proc=2.6.1=h7f98852_0
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
libxml2=2.9.12=h72842e0_0
llvmlite=0.36.0=py37h9d7f4d0_0
locket=0.2.0=py_2
lz4-c=1.9.3=h295c915_1
markdown=3.3.5=pyhd8ed1ab_0
markupsafe=2.0.1=py37h5e8e339_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
msgpack-python=1.0.2=py37h2527ec5_1
multidict=5.1.0=py37h27cfd23_2
nccl=2.10.3.1=h1a5f58c_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
ninja=1.10.2=hff7bd54_1
nose=1.3.7=pyhd3eb1b0_1006
numba=0.53.1=py37hb11d6e1_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
nvtx=0.2.3=py37h5e8e339_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=py37_0
openh264=2.1.0=hd408876_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
orc=1.6.7=h89a63ab_2
packaging=21.0=pyhd8ed1ab_0
pandas=1.2.4=py37h219a48f_0
parquet-cpp=1.5.1=2
partd=1.2.0=pyhd8ed1ab_0
pillow=8.3.1=py37h2c7a002_0
pip=21.0.1=py37h06a4308_0
protobuf=3.16.0=py37hcd2ae1e_0
psutil=5.8.0=py37h5e8e339_1
pyarrow=1.0.1=py37hb63ea2f_40_cuda
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pydeprecate=0.3.1=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_0
pymongo=3.12.0=py37h295c915_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py37h89c1867_4
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch=1.8.2=py3.7_cuda10.2_cudnn7.6.5_0
pytorch-lightning=1.5.2=pyhd8ed1ab_0
pytz=2021.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=5.4.1=py37h5e8e339_0
re2=2021.04.01=h9c3ff4c_0
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rmm=0.19.0=cuda_10.2_py37_g7065af3_0
rsa=4.7.2=pyh44b312d_0
s2n=1.0.10=h9b69904_0
scikit-learn=0.24.2=py37ha9443f7_0
scipy=1.7.0=py37h29e03ee_1
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
snappy=1.1.8=he1b5a44_3
sortedcontainers=2.4.0=pyhd8ed1ab_0
spdlog=1.7.0=hc9558a2_2
sqlite=3.36.0=hc218d9a_0
tblib=1.7.0=pyhd8ed1ab_0
tensorboard=2.7.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py37h7f0c10b_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
torchaudio=0.8.2=py37
torchmetrics=0.6.0=pyhd8ed1ab_0
torchvision=0.9.2=py37_cu102
tornado=6.1=py37h5e8e339_1
tqdm=4.62.3=pyhd8ed1ab_0
treelite=1.1.0=py37hc731546_0
treelite-runtime=1.1.0=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
ucx=1.9.0+gcd9efd3=cuda10.2_0
ucx-proc=1.0.0=gpu
ucx-py=0.19.0=py37_gcd9efd3_0
urllib3=1.26.7=pyhd8ed1ab_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd3eb1b0_1
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py37h5e8e339_2
zict=2.0.0=py_0
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
